refactor(cli): extract frequency solve session module
Move per-frequency solve logic out of main.rs into a new
apps/nec-cli/src/solve_session.rs module:

- structs: FrequencySolveResult, SweepPointSummary,
  PulseCurrentSourceConstraint, HybridLanePlan
- math helpers: l2_norm, matrix_diagonal_spread,
  residual_zi_minus_v, residual_hallen
- pulse helpers: collect/apply pulse current source constraints,
  pulse_current_source_voltage
- report builders: build_feedpoint_rows, build_source_rows,
  build_load_rows
- frequency/dispatch: frequencies_from_fr, build_hybrid_lane_plan
- main function: solve_frequency_point (new explicit
  sinusoidal_topology_supported: bool param, computed once in
  main() before the solve closure)

main.rs now contains only frontend wiring, geometry validation
helpers, and warning functions.
